requirements.txt
numpy<2
torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
pyg-lib==0.4.0+pt23cu121
torch-scatter==2.1.2+pt23cu121
torch-sparse==0.6.18+pt23cu121
torch-cluster==1.6.3+pt23cu121
torch-spline-conv==1.2.2+pt23cu121
torch-geometric==2.6.1
ansys-mapdl-core
ansys-tools-visualization_interface
pyvistal
